JSON のファイル読み書きのメモリサイズ制限を緩和 事実上メモリがある限り読み書き可能にした

+++ Cores.NET/Cores.Basic/Common/Base/Buffer.cs
@@ -2237,7 +2237,7 @@ namespace IPA.Cores.Basic
                     }
                 }
 
-                Debug.Assert(ret.Select(x => x.Size).Sum() == size);
+                Debug.Assert(ret.Select(x => (long)x.Size).Sum() == size);
 
                 readSize = size;
